Postagram_project/postagram_backend_django/requirements.txt
Django==5.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
django-cors-headers==4.3.1
pytest-django==4.8.0